+++ Model/Attribute/InputType/Select.php
@@ -151,11 +151,19 @@ class Select extends AbstractInputType
      */
     public function getDefaultValue()
     {
-        $defaultValue = parent::getDefaultValue();
-        if ($defaultValue) {
+        $defaultValue = (string) parent::getDefaultValue();
+        if ($defaultValue !== '') {
             return explode(',', $defaultValue);
         }
 
         return false;
+    }
+
+    /**
+     * @return true
+     */
+    public function hasEmptyOption()
+    {
+        return true;
     }
 }